==> Code/03_Baseline_Model.ipynb <==
#import all library
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.metrics import mean_absolute_error,mean_squared_error

pd.set_option('display.max_columns', 100)
pd.set_option('display.max_rows', 100)

# read csv file
housing = pd.read_csv('../data/cleaned_train_data_wo_lat_long.csv')
housing.head()

avg_resales_price = housing['resale_price'].mean()
print("average resale price:", avg_resales_price)

y_pred = [avg_resales_price]*len(housing)
y = housing['resale_price']
mse = mean_squared_error(y,y_pred)
mae = mean_absolute_error(y,y_pred)
rmse = mean_squared_error(y, y_pred, squared = False)

print("MSE:", mse)
print("MAE:", mae)
print("rmse:", rmse)
