==> Code/04_Preprocessing_and_Feature_Engineering.ipynb <==
#import all library
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
#import library for VIF
#from statsmodels.stats.outliers_influence import variance_inflation_factor

pd.set_option('display.max_columns', 100)
pd.set_option('display.max_rows', 100)

# read csv file
housing = pd.read_csv('../data/cleaned_combined_data_wo_lat_long.csv')
housing.head()
housing.shape

### Feature_Engineering

# creating another feature - diff_from_max_lvl
#housing['diff_from_max_lvl'] = housing['max_floor_lvl'] - housing['mid_storey']

# creating another feature - percent_rental_flat
housing['percent_rental_flat'] = (housing['1room_rental'] + housing['2room_rental'] + housing['3room_rental'] + housing['other_room_rental'])/housing['total_dwelling_units']


# creating another feature - percent_5_room_and_above
housing['percent_4room_and_above_flat'] = (housing['4room_sold'] + housing['5room_sold'] + housing['exec_sold'] + housing['multigen_sold'] + housing['studio_apartment_sold'])/housing['total_dwelling_units']


rooms = ['3 ROOM', '4 ROOM', '5 ROOM', 'EXECUTIVE','MUTI_GENERATION']

# creating another feature - If the resale unit is 3 room-flat and above, and there are 1 room-flat or 2 room-flat, then 1 else 0
housing['present_of_3room_and_below'] = (
    (housing['flat_type'].isin(rooms)) & 
    ((housing['1room_sold'] + housing['2room_sold'] + housing['3room_sold']) > 0)
).astype(int)

# Interaction Term 
housing['distance_for_hawker_and_mrt'] = housing['hawker_nearest_distance'] * housing['mrt_nearest_distance']
housing['distance_for_mall_and_mrt'] = housing['mall_nearest_distance'] * housing['mrt_nearest_distance']



Dropping columns that is not going to be used in the regression model 

cols_to_drop = ['tranc_yearmonth']
cols_for_corr_test = housing.drop(columns = cols_to_drop , inplace = False)


cols_for_corr_test.corr()['resale_price'].sort_values()

cols_for_corr_test['percent_4room_and_above_flat'].corr(cols_for_corr_test['present_of_3room_and_below'])

### Drop Columns

Dropping columns that might have high multicollinearity and variables used to create interaction term 
- region, storey_range, full_flat_type and etc
- mall_nearest_distance, mrt_nearest_distance,hawker_nearest_distance

cols_to_drop = ['full_flat_type','present_of_3room_and_below','bus_stop_name','pri_sch_name','tranc_year','tranc_month','mall_nearest_distance','mrt_nearest_distance','hawker_nearest_distance','region','storey_range','1room_sold','2room_sold','3room_sold','4room_sold','5room_sold','exec_sold','multigen_sold','studio_apartment_sold','1room_rental','2room_rental','3room_rental','other_room_rental','sec_sch_name']
housing2 = housing.drop(columns= cols_to_drop ,inplace=False)
housing2.head()

housing2.shape




#def calc_vif(X):

    # Calculating VIF
    #vif = pd.DataFrame()
    #vif["variables"] = X.columns
    #vif["VIF"] = [variance_inflation_factor(X.values, i) for i in range(X.shape[1])]

    #return(vif)

#X = housing2[housing2['train']==1].drop(columns=['flat_type','flat_model','planning_area','mrt_name'])

#calc_vif(X)

# Mutual information on Numerical Input
#from sklearn.feature_selection import mutual_info_regression

#y = housing2[housing2['train']==1]['resale_price']


#mi_scores = mutual_info_regression(X,y)
#mi_scores = pd.Series(mi_scores, name="MI Scores", index = X.columns)
#mi_scores = mi_scores.sort_values(ascending=False)

#mi_scores

### Creating dummy variables for categories variables

####  flat_type, flat_model, full_flat_type, planning_area, mrt_name, bus_stop_name, pri_sch_name,region, tranc_year, tranc_month, tranc_yearmonth 

#dummify_list = ['flat_type','flat_model','full_flat_type','planning_area','mrt_name','bus_stop_name','pri_sch_name','region','tranc_year','tranc_month','tranc_yearmonth']
dummify_list = ['flat_type','flat_model','mrt_name','planning_area','tranc_yearmonth'] #'mrt_name'

for column in dummify_list:
    # Create dummy variables for each column and concatenate them to the housing2 DataFrame
    dummy_variables = pd.get_dummies(housing[column], prefix=column, drop_first=True)
    housing2 = pd.concat([housing2, dummy_variables], axis=1)

# Drop the original categorical columns from the housing2 DataFrame
housing2.drop(columns=dummify_list, inplace=True)

housing2.head()

#Save to csv file
# Drop resale price outliner as it only contribute to 3% - from 02_EDA finding
#train = housing2[((housing2['train']==1) & (housing2['resale_price'] < 781888))]

train = housing2[housing2['train']==1]
test = housing2[housing2['train']==0]

train.to_csv('../data/regression_train.csv', index=False)

test.to_csv('../data/regression_test.csv', index=False)

housing2[housing2['train']==0]
