==> Code/06_Ridge_And_Lasso_Model.ipynb <==
#import all library
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

import sklearn

from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler,PowerTransformer,Normalizer
from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.metrics import mean_absolute_error,mean_squared_error,r2_score

from sklearn.linear_model import Ridge
from sklearn.linear_model import RidgeCV
from sklearn.linear_model import Lasso, LassoCV
from sklearn.linear_model import ElasticNet
pd.set_option('display.max_columns', 100)
pd.set_option('display.max_rows', 100)

#Read CSV file
housing = pd.read_csv('../Data/regression_train.csv')
housing = housing[housing['resale_price'] < 1000000] 
housing.head()                 

X=housing.drop(columns=['resale_price','indicator_to_drop','train','vacancy','id'],inplace=False)
y=housing['resale_price']

X_train, X_test, y_train, y_test = train_test_split(X,
                                                    y,
                                                    test_size = 0.2,
                                                    random_state = 42)

sklearn.metrics.SCORERS.keys()

# Set up a list of ridge alphas to check.
# np.logspace generates 100 values equally between 0 and 5,
# then converts them to alphas between 10^0 and 10^5.
r_alphas = np.logspace(0, 10, 1)

# Cross-validate over our list of ridge alphas.
ridge_cv = RidgeCV(alphas=r_alphas, scoring='max_error', cv=3)

# Fit model using best ridge alpha!
ridge_cv = ridge_cv.fit(X_train, y_train)

# Here is the optimal value of alpha
ridge_cv.alpha_

# Instantiate a StandardScaler + ridge pipeline.
pipe_r = Pipeline([
    ('ps', StandardScaler()),
    ('r', Ridge(alpha=1))
])

def regression_metrices(pipe):
    
    pipe.fit(X_train,y_train)
    
    print('r-score on train dataset:',pipe.score(X_train,y_train))
    print('r-score on test dataset:', pipe.score(X_test,y_test))
    
    y_pred = pipe.predict(X_test)
    
    # Evaluation metrics for the actual model
    mae = mean_absolute_error(y_test, y_pred)
    mse = mean_squared_error(y_test, y_pred)
    rmse = mean_squared_error(y_test, y_pred, squared = False)

    print("MAE:", mae)
    print("MSE:", mse)
    print("RMSE:", rmse)

regression_metrices(pipe_r)

### Lasso

# Set up a list of ridge alphas to check.
# np.logspace generates 100 values equally between 0 and 5,
# then converts them to alphas between 10^0 and 10^5.
l_alphas = np.logspace(-3, 0, 10)

# Cross-validate over our list of ridge alphas.
lasso_cv = LassoCV(alphas=l_alphas, cv=5, max_iter=10000)

# Fit model using best ridge alpha!
lasso_cv = lasso_cv.fit(X_train, y_train)

# Here is the optimal value of alpha
lasso_cv.alpha_

# Instantiate a StandardScaler + ridge pipeline.
pipe_l = Pipeline([
    ('ss', StandardScaler()),
    ('l', Lasso(alpha=0.21544346900318823))
])

regression_metrices(pipe_l)

### ElasticNet

# Instantiate a StandardScaler + ridge pipeline.
pipe_e = Pipeline([
    ('e', ElasticNet(alpha=0.5))
])

regression_metrices(pipe_e)
