==> Code/10_Kaggle_Submissions-Final-LogY.ipynb <==
#import all library
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

import sklearn

from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler,PowerTransformer,Normalizer
from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.metrics import mean_absolute_error,mean_squared_error,r2_score

from sklearn.linear_model import Ridge
from sklearn.linear_model import RidgeCV

pd.set_option('display.max_columns', 220)
pd.set_option('display.max_rows', 220)

#Read CSV file
housing = pd.read_csv('../Data/regression_train.csv')
housing = housing[housing['resale_price'] < 1000000] #781888
housing.shape

housing.head()   

cols_to_drop_list = ['flat_model_Maisonette',
 'flat_model_Apartment',
 'mrt_name_Eunos',
 'mrt_name_Chinese Garden',
 'mrt_name_Little India',
 'cutoff_point_rank',
 'mrt_name_Bukit Gombak',
 'mrt_name_Serangoon',
'resale_price','indicator_to_drop','train','vacancy','id'] # Based on OLS result in 06, P value less than 0.05

X=housing.drop(columns= cols_to_drop_list ,inplace=False)
y=housing['resale_price']

X_train, X_test, y_train, y_test = train_test_split(X,
                                                    np.log(y),
                                                    test_size = 0.2,
                                                    random_state = 42)

X_train.shape

# Instantiate.
ss = StandardScaler()

# Fit and transform.
X_train_ss = ss.fit_transform(X_train)

# Transform.
X_test_ss = ss.transform(X_test)

r = Ridge(alpha=1)
r.fit(X_train_ss, y_train);


print('r-score on train dataset:',r.score(X_train_ss,y_train))
print('r-score on test dataset:', r.score(X_test_ss,y_test))
    
y_pred = r.predict(X_test_ss)
    
# Evaluation metrics for the actual model
mae = mean_absolute_error(y_test, y_pred)
mse = mean_squared_error(y_test, y_pred)
rmse = mean_squared_error(y_test, y_pred, squared = False)

print("MAE:", mae)
print("MSE:", mse)
print("RMSE:", rmse)

### Test.csv

#Read CSV file
housing_test = pd.read_csv('../Data/regression_test.csv')   

housing_test.shape

X=housing_test.drop(columns=cols_to_drop_list,inplace=False)
X_ss = ss.transform(X)

X_ss.shape

print(X.isnull().sum())

X.dropna(inplace=True)

X.shape

y_pred = r.predict(X_ss)

y_pred

result = pd.DataFrame(y_pred).reset_index()

result = housing_test[['id']]
result['predicted'] = np.exp(y_pred)

result

result.to_csv('../data/test_result.csv', index=False)
